% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{convertEpstein}
\alias{convertEpstein}
\title{delta 18 O convert
Convert suite of delta 18 O and convert it to temperature}
\usage{
convertEpstein(d)
}
\arguments{
\item{d}{a real or vector of real representing values of delta 18 0}
}
\description{
delta 18 O convert
Convert suite of delta 18 O and convert it to temperature
}
