% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{countStrategies}
\alias{countStrategies}
\title{countStrategies}
\usage{
countStrategies(pop, lowerlim = 0.25, upperlim = 0.75)
}
\arguments{
\item{lowerlim}{the maximum value for the other gene (y or z) to be considered as "pure" learning}

\item{upperlim}{the minimum value for the gene (y or z) to be considered as "pure" learning}
}
\description{
A function to count the number of pure social learning and pure individual learner
}
