% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{environment}
\alias{environment}
\title{Environment generator}
\usage{
environment(N, omega, delta, vt = NULL)
}
\arguments{
\item{N}{the length of the timeserie to generate}

\item{omega}{the slope of the power distribution (called omega in whitehead)}

\item{delta}{the standard deviation of the global environmental fluctuation}

\item{vt}{if not NULL, the  mean of the optimum increase at a rate vt}
}
\value{
a list of N optima theta
}
\description{
return a list of opitmum as a "1/f-noise" of N steps with sd = delta and the slope of it's spectrum decompoistion = -omega + possibility to increase the mean of the environment at a rate vt (in wich case the slop may not be -omega)
}
