% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corefunctions.R
\name{generatePop}
\alias{generatePop}
\title{Generate population}
\usage{
generatePop(n, distrib, df = F)
}
\arguments{
\item{n}{number of agents in the population}

\item{distrib}{should be a list with 3 elements x yz)}

\item{df}{if TRUE return a data.frame (easier to handle but much much slower, default is FALSE)}
}
\description{
Function that generates a population
}
