% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{getClosest}
\alias{getClosest}
\title{Resample using closest data
create an evenly spaced vector of date and associate to it the closest datapoint from the original dataset}
\usage{
getClosest(data, year, by)
}
\arguments{
\item{data}{a vector of measure}

\item{year}{a vector of date}

\item{by}{the ne sampling rate in year}
}
\value{
a two columns data frame. Each column is of dimension: seq(min(year),max(year),by=by), one with the years and other with associated theta
}
\description{
Resample using closest data
create an evenly spaced vector of date and associate to it the closest datapoint from the original dataset
}
