% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{getDateResolution}
\alias{getDateResolution}
\title{Get Lowest Resolution
get the lowest resolution of the vector years}
\usage{
getDateResolution(years)
}
\arguments{
\item{years}{a vector of years}
}
\description{
Get Lowest Resolution
get the lowest resolution of the vector years
}
