% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{getLogSpec}
\alias{getLogSpec}
\title{Get the log10 of the spectrum density and frequencies of a vector}
\usage{
getLogSpec(data, year)
}
\description{
Get the log10 of the spectrum density and frequencies of a vector
}
