% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{getMean2}
\alias{getMean2}
\title{Resample using mean}
\usage{
getMean2(data, year, by)
}
\arguments{
\item{data}{a vector of measure}

\item{year}{a vector of date}

\item{by}{the ne sampling rate in year}
}
\value{
a two columns data frame. Each column is of dimension: seq(min(year),max(year),by=by), one with the years and other with associated theta
}
\description{
create an evenly spaced vector of date and associate to it the mean of the datapoints falling within the new intervals from the original dataset
}
