% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{getOmega}
\alias{getOmega}
\title{return the coefficient of a linear fit to the spectrum decomposition of a time serie t}
\usage{
getOmega(t)
}
\description{
return the coefficient of a linear fit to the spectrum decomposition of a time serie t
}
