% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{getOmega2}
\alias{getOmega2}
\title{Get the absolute slope of a linear fit between two vectors, possibility to limit the fit to a region of the data (using u and sp)}
\usage{
getOmega2(x, y, u = NULL, sp = 0.5)
}
\description{
Get the absolute slope of a linear fit between two vectors, possibility to limit the fit to a region of the data (using u and sp)
}
