% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{getOmegaWrap}
\alias{getOmegaWrap}
\title{getOmegaWrap return the slope of the spectrum decomposition of a vector}
\usage{
getOmegaWrap(data)
}
\arguments{
\item{data}{a vector}
}
\value{
a double
}
\description{
getOmegaWrap return the slope of the spectrum decomposition of a vector
}
