% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{getRgbScale}
\alias{getRgbScale}
\title{Plot single run}
\usage{
getRgbScale(results)
}
\arguments{
\item{results}{the output of evosolearn()}
}
\description{
RGB Scale
}
\details{
Function that gives the rgb scale of a results
}
