% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{getSpectrumMTM}
\alias{getSpectrumMTM}
\title{Get spectrum decomposition}
\usage{
getSpectrumMTM(data, freq, nw = 2, k = 3, ...)
}
\arguments{
\item{data}{a vector of measure}

\item{freq}{a vector of time at which each measure of \emph{data} has been done}
}
\description{
wrapper to calculate the spectrum decomposition of a timeserie using Multitaper Methods
}
