% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getSummary}
\alias{getSummary}
\title{get some useful summary stat from a model run}
\usage{
getSummary(fullmat, nstep, sumstat = mean, vars)
}
\arguments{
\item{fullmat}{a matrice with the result of a simulation}

\item{nstep}{how many timestep should we use to compute the summary statitistic}

\item{sumstat}{function used to compute the summary statistic}

\item{vars}{a vector with the variable we want to summarize using sumstat}
}
\description{
get some useful summary stat from a model run
}
