% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{getVarXMeanW}
\alias{getVarXMeanW}
\title{get some useful summary stat from model run}
\usage{
getVarXMeanW(n, tstep, omega, delta, b, K, mu, E, sigma, pop, m, gene,
  nstep, sumstat)
}
\arguments{
\item{nstep}{how many timestep should we use to compute the summary statitistic}

\item{sumstat}{function used to compute the summary statistic}
}
\description{
get some useful summary stat from model run
}
