% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{getWindows}
\alias{getWindows}
\title{apply the function f on subsets of size w of the vector data}
\usage{
getWindows(data, w, f)
}
\description{
apply the function f on subsets of size w of the vector data
}
