% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{interpolate}
\alias{interpolate}
\title{Function interpolate}
\usage{
interpolate(theta, times, finalres, delta = 0, omega = 0,
  delta2 = NULL)
}
\arguments{
\item{theta:}{the original theta vector to complete}

\item{times:}{the time associated to the thetas}

\item{finalres:}{the resolution, in the unit used in the vector times, at which new data had to be generate. if times is in year and finalres is 1, the vector thetat will correspond to data ofr each eyar, if time is in thousand of years (as in LR04 stack) and finalres is 0.5 thus the output will correspond to one point every 500 years.}

\item{omega:}{autocorrelation of the noise use to generate interpolate datapoints}

\item{delta:}{variance of the noise used for bigger gap}

\item{delta2:}{variance of the noise used for smaller gaps}
}
\value{
a two columns data frame. Each column is a of dimension: seq(1,length(times),by=finalres), one with the years and other with associated theta
}
\description{
A function to add intermediate values between two optimums theta
}
