% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{martrat}
\alias{martrat}
\title{Martrat dataset}
\format{A data frame with 3311 rows and 4 variables}
\usage{
data(martrat)
}
\description{
A dataset containing detla 18 0 and rescaled delta 18 0 from martrat
}
\details{
\itemize{
  \item year. date of measurment
  \item SST sea surface temperature
  \item dTsVscales rescalede variation 
}
}
\keyword{datasets}
