% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ngrip2}
\alias{ngrip2}
\title{NGRIP2 dataset}
\format{A data frame with 6113 rows and 3 variables}
\usage{
data(ngrip2)
}
\description{
A dataset containing detla 18 0 and rescaled delta 18 0 from martrat
}
\details{
\itemize{
  \item year. date of measurment
  \item d0ice  delta 0 18 from ice
  \item dTsVscales rescaled variation 
  \item resampled years 
}
}
\keyword{datasets}
