% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plotDistThetaNW}
\alias{plotDistThetaNW}
\title{Plot fitness and distance to optimum}
\usage{
plotDistThetaNW(results, ...)
}
\arguments{
\item{restults}{the output of evosolearn()}
}
\description{
Function that plot mean a standard deviation
}
