% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plotThetaPhenotypes}
\alias{plotThetaPhenotypes}
\title{Plot phenotype and  optimum}
\usage{
plotThetaPhenotypes(results, ...)
}
\arguments{
\item{restults}{the output of evosolearn()}
}
\description{
Function that plot mean phenotype a standard deviation
}
