% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corefunctions.R
\name{reproduction}
\alias{reproduction}
\title{Reproduction process}
\usage{
reproduction(pop, selected, nchilds)
}
\arguments{
\item{pop}{the current population with characteristics of all individual}

\item{selected}{a vector with the id of the selected individuals}

\item{nchilds}{a vector with the number of children for each individual}
}
\value{
a matrix qiht the list of children
}
\description{
Function that generate a new population given a list of selected individuals
}
