% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corefunctions.R, R/protomodels.R
\name{selection}
\alias{selection}
\title{Selection function}
\usage{
selection(w, b, n, K)

selection(w, b, n, K)
}
\arguments{
\item{w:vector}{with fitness}

\item{b:numeric}{value for the rate of birth}

\item{n:numeric}{value for the size of the pop (redundant/should be equal to length(w))}

\item{K:environment}{carrying capacity}

\item{w:vector}{with fitness}

\item{b:numeric}{value for the rate of birth}

\item{n:numeric}{value for the size of the pop (redundant/should be equal to length(w))}

\item{K:environment}{carrying capacity}
}
\value{
: numeric vector of probabilities

: numeric vector of probabilities
}
\description{
Selection function

Selection function
}
