% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corefunctions.R
\name{socialLearning}
\alias{socialLearning}
\title{Social Learning}
\usage{
socialLearning(newpop, reference, thetat = NULL, sls = "random")
}
\arguments{
\item{newpop}{a dataframe with fitness and agents ID}

\item{reference}{a dataframe with phenotype and agents ID}

\item{sls}{a string define the sls of copy to be done: in "parents","best","average","randon"}

\item{theta}{the value of the optimum}
}
\value{
: a unique numeric value or a vector of size nrow(newpop) with phenotypes to be copied
}
\description{
Function that select a new phenotype within a given reference group and a certain strategy
}
