% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{TK95}
\alias{TK95}
\title{From tuneR package}
\usage{
TK95(N, alpha = 1)
}
\arguments{
\item{alpha:}{the slope of the power distribution (called omega in whitehead)}

\item{N:}{the length of the timeserie to generate}
}
\description{
Based on Timmer and Koening 1995: https://ui.adsabs.harvard.edu/abs/1995A\%26A...300..707T/abstract
}
