% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corefunctions.R
\name{updateOutput}
\alias{updateOutput}
\title{Update output}
\usage{
updateOutput(output = NULL, pop, statfun, statvar)
}
\arguments{
\item{pop}{the current population upon wihch statistics have to be calculated}

\item{statfun}{a vector with the different function we apply on the population}

\item{statvar}{a vector with the different varaible we measure in the population}

\item{ouptut}{a current output to be update or initialized if NULL}
}
\description{
Function that updates the matrix stored that willbe returned at the end of the simulation.
}
