% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corefunctions.R, R/protomodels.R
\name{updateOutputLine}
\alias{updateOutputLine}
\title{updateOutputLine}
\usage{
updateOutputLine(pop, statfun, statvar, getname = F, prop = T)

updateOutputLine(pop, statfun, statvar, getname = F, prop = T)
}
\arguments{
\item{pop}{the current population upon wihch statistics have to be calculated}

\item{statfun}{a vector with the different function we apply on the population}

\item{statvar}{a vector with the different varaible we measure in the population}

\item{prop}{if the proporition of different strategies should be ouptut}

\item{ouptut}{a current output to be update or initialized if NULL}

\item{ouptut}{a current output to be update or initialized if NULL}

\item{pop}{the current population upon wihch statistics have to be calculated}

\item{statfun}{a vector with the different function we apply on the population}

\item{statvar}{a vector with the different varaible we measure in the population}

\item{prop}{if the proporition of different strategies should be ouptut}
}
\description{
updateOutputLine

updateOutputLine
}
