% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vostok}
\alias{vostok}
\title{Vostok dataset}
\format{A data frame with 3311 rows and 4 variables}
\usage{
data(vostok)
}
\description{
A dataset containing detla 18 0 and rescaled delta 18 0 from vostok
}
\details{
\itemize{
  \item year. date of measurment
  \item dD  delta D
  \item dTsVscales rescalede variation 
  \item resampled years 
}
}
\keyword{datasets}
