% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corefunctions.R
\name{evosolearn}
\alias{evosolearn}
\title{Main function}
\usage{
evosolearn(n, tstep, E = c(x = 0.01, y = 0.01, z = 0.01), sigma = c(s =
  1, y = 1, z = 1), omega, delta, b, K, mu = c(x = 0.3, y = 0.3, z =
  0.3), genes = c("x", "y", "z"), m = c(x = 0.3, y = 0.3, z = 0.3),
  sls = "best", log = F, pop = NULL, allpops = F,
  statfun = c("mean", "var"), statvar = c("x", "y", "z", "gp", "ilp",
  "p", "w"), outputrate = 1, vt = NULL, theta = NULL, prop = TRUE,
  repro = "asex", selection = T)
}
\arguments{
\item{pop}{the current population upon wihch statistics have to be calculated}

\item{statfun}{a vector with the different function we apply on the population}

\item{statvar}{a vector with the different varaible we measure in the population}

\item{prop}{if the proporition of different strategies should be ouptut}

\item{ouptut}{a current output to be update or initialized if NULL}
}
\description{
Main function
}
