% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plotResults}
\alias{plotResults}
\title{Plot all}
\usage{
plotResults(results, statfun = c("mean", "sd"), statvar = c("w", "p",
  "x", "y", "z"), multi = F, N = T, theta = T, addrgb = T)
}
\arguments{
\item{restults}{the output of evosolearn()}
}
\description{
Function that plot mean and standard for all values
}
