% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corefunctions.R, R/protomodels.R
\name{socialLearning}
\alias{socialLearning}
\title{Social Learning, a function defining how individual socially copies the others}
\usage{
socialLearning(newpop, reference, thetat = NULL, sls = "random")

socialLearning(newpop, reference, thetat = NULL, sls = "random")
}
\arguments{
\item{newpop:}{a dataframe with fitness and agents ID}

\item{reference:}{a dataframe with phenotype and agents ID}

\item{sls:}{a string define the sls of copy to be done: \in {"parents","best","average","randon"}}

\item{newpop:}{a dataframe with fitness and agents ID}

\item{reference:}{a dataframe with phenotype and agents ID}

\item{sls:}{a string define the sls of copy to be done: \in {"parents","best","average","randon"}}
}
\value{
: a unique numeric value or a vector of size nrow(newpop) with phenotypes to be copied

: a unique numeric value or a vector of size nrow(newpop) with phenotypes to be copied
}
\description{
Social Learning, a function defining how individual socially copies the others
}
